requirements.txt
-e .

pytest
pytest-cov
flake8
autoflake
pep8
autopep8

backports.tempfile
tensorflow